Enhance Manga model with JSON serialization; update favorite management to associate with user accounts and improve UI in intro pages

+++ comic_book_store/lib/controllers/favouriteControler.dart
@@ -1,11 +1,12 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:comic_book_store/api/jikan_service.dart';
 import 'package:comic_book_store/models/manga.dart';
-
-import '../api/jikan_service.dart';
+import 'package:firebase_auth/firebase_auth.dart';
 
 class MangaController {
   final JikanService _jikanService = JikanService();
   final FirebaseFirestore _firestore = FirebaseFirestore.instance;
+  final FirebaseAuth _auth = FirebaseAuth.instance;
 
   Future<Manga?> fetchMangaDetails(int mangaId) async {
     try {
@@ -19,8 +20,13 @@ class MangaController {
 
   Future<bool> isFavorite(int mangaId) async {
     try {
+      final user = _auth.currentUser;
+      if (user == null) return false;
+
       final favoriteDoc = await _firestore
           .collection('favorites')
+          .doc(user.email)
+          .collection('userFavorites')
           .doc(mangaId.toString())
           .get();
       return favoriteDoc.exists;
@@ -32,18 +38,25 @@ class MangaController {
 
   Future<void> toggleFavorite(Manga manga, bool isFavorite) async {
     try {
-      final favoriteRef =
-          _firestore.collection('favorites').doc(manga.malId.toString());
+      final user = _auth.currentUser;
+      if (user == null) return;
+
+      final favoriteRef = _firestore
+          .collection('favorites')
+          .doc(user.email)
+          .collection('userFavorites')
+          .doc(manga.malId.toString());
 
       if (isFavorite) {
         // Remove from favorites
         await favoriteRef.delete();
       } else {
         // Add to favorites
-        // await favoriteRef.set({
-        //   ...manga.toJson(),
-        //   'addedAt': FieldValue.serverTimestamp(),
-        // });
+        await favoriteRef.set({
+          'userEmail': user.email,
+          ...manga.toJson(),
+          'addedAt': FieldValue.serverTimestamp(),
+        });
       }
     } catch (e) {
       print('Error toggling favorite: $e');
